Added entities validation on select.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1283 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 13ba5e3cde06dda079d499db428389a40d8cf44e

+++ Source/Mapper.cpp
@@ -4520,13 +4520,18 @@ void FOMapper::SScriptFunc::Global_DeleteEntities( CScriptArray* entities )
     for( int i = 0, j = entities->GetSize(); i < j; i++ )
     {
         Entity* entity = *(Entity**) entities->At( i );
-        if( entity )
+        if( entity && !entity->IsDestroyed )
             Self->DeleteEntity( entity );
     }
 }
 
 void FOMapper::SScriptFunc::Global_SelectEntity( Entity* entity, bool set )
 {
+    if( !entity )
+        SCRIPT_ERROR_R( "Entity arg is null." );
+    if( entity->IsDestroyed )
+        SCRIPT_ERROR_R( "Entity arg is destroyed." );
+
     if( set )
         Self->SelectAdd( entity );
     else
@@ -4540,6 +4545,9 @@ void FOMapper::SScriptFunc::Global_SelectEntities( CScriptArray* entities, bool 
         Entity* entity = *(Entity**) entities->At( i );
         if( entity )
         {
+            if( entity->IsDestroyed )
+                SCRIPT_ERROR_R( "Entity in array is destroyed." );
+
             if( set )
                 Self->SelectAdd( entity );
             else
